+++ apis/cassandra/v1beta1/cassandradatacenter_types.go
@@ -77,6 +77,10 @@ const (
 	// UseClientBuilderAnnotation enforces the usage of new config builder from k8ssandra-client for versions that would otherwise use the cass-config-builder
 	UseClientBuilderAnnotation = "cassandra.datastax.com/use-new-config-builder"
 
+	// TrackCleanupTasksAnnotation enforces the operator to track cleanup tasks after doing scale up. This prevents other operations to take place until the cleanup
+	// task has completed.
+	TrackCleanupTasksAnnotation = "cassandra.datastax.com/track-cleanup-tasks"
+
 	AllowUpdateAlways AllowUpdateType = "always"
 	AllowUpdateOnce   AllowUpdateType = "once"
 
@@ -561,6 +565,25 @@ func (status *CassandraDatacenterStatus) GetConditionStatus(conditionType Datace
 		}
 	}
 	return corev1.ConditionUnknown
+}
+
+func (status *CassandraDatacenterStatus) AddTaskToTrack(objectMeta metav1.ObjectMeta) {
+	if status.TrackedTasks == nil {
+		status.TrackedTasks = make([]corev1.ObjectReference, 0, 1)
+	}
+
+	status.TrackedTasks = append(status.TrackedTasks, corev1.ObjectReference{
+		Name:      objectMeta.Name,
+		Namespace: objectMeta.Namespace,
+	})
+}
+
+func (status *CassandraDatacenterStatus) RemoveTrackedTask(objectMeta metav1.ObjectMeta) {
+	for index, task := range status.TrackedTasks {
+		if task.Name == objectMeta.Name && task.Namespace == objectMeta.Namespace {
+			status.TrackedTasks = append(status.TrackedTasks[:index], status.TrackedTasks[index+1:]...)
+		}
+	}
 }
 
 func (dc *CassandraDatacenter) GetConditionStatus(conditionType DatacenterConditionType) corev1.ConditionStatus {